requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
pyyaml==6.0.1
pytest==7.4.3
black==23.12.1
flake8==6.1.0
statsmodels==0.14.4
